requirements.txt
snowflake-cli
snowflake-cli-labs